C prototipo/Sprint 4/app web/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
chart.js==4.0.1
